CRUD for client goals etc is done(ish)
Most CRUD operations for serving the functionality that allow you to set goals for clients, and perform assignments and measurements, are finished. There are a few issues still (the delete requests are not all working as expected), but I'm going to focus on front end now.

Also updated the signup and login, password hashing is now performed